Phew! Moved commands to separate files
