+++ src/shard/commands/qtprefix.ts
@@ -1,0 +1,11 @@
+import { CmdFn } from ".";
+import { setPrefix } from "../../subs";
+import { translated } from "../post";
+
+const qtprefix: CmdFn = async ({ args }, qChannel) => {
+    const prefix = args.shift();
+    setPrefix(qChannel.guildId(), prefix);
+    translated(qChannel, 'prefixSuccess', { prefix });
+};
+
+export default qtprefix
