+++ src/shard/discord/index.d.ts
